backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "cryptography>=46.0.3",
    "dotenv>=0.9.9",
    "fastapi>=0.128.0",
    "langchain-chroma>=0.1.2",
    "langchain-openai>=1.1.7",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-extra-types>=2.11.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.45",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
]
